Adding App Icon.
Modification of Favorites and History Storyboard.
TableViewModel: Adding SearchBar and filters. Renaming functions to prepare implementation of HistoryViewModel for the History Scene. Deleting function.
Creation of HistoryViewModel and updating of HistoryViewController based on FavoritesViewModel and TableViewModel.
PlaceManager: including history: renaming of some functions, creation of functions for Favorite and History (findPlaceFavorite, findPlaceHistory, updateHistory and removeFromHistory). Adding history in addPlace function and "Place" object