PlaceManager: rewriting the deleteAll function and creating a findPlaceFavoriteOrHistory function to better manage object deletion and keep favorites and histories. deleteAll: All objects are deleted except favorites and histories. findPlaceFavoriteOrHistory: all objects are added to the database unless already present because favorites or history.
NetworkModel : Removal of the old favorites and history saving system and implementation of findPlaceFavoriteOrHistory in the 4 functions: museum, theatre, garden and library and reorganization of the file accordingly.

Adaptation of the library function after rewriting some LibraryDecodable data.

removeFromHistory, updateHistory and updateFavorite: updated by adding coordinates to avoid duplicates in favorites and history.